+++ bulletin_board/server/app/services/voting_scheme/dummy/bulletin_board.rb
@@ -18,7 +18,7 @@ module VotingScheme
     class BulletinBoard < VotingScheme::BulletinBoard
       include Dummy
 
-      RESULTS = ["tally.share", "end_tally"].freeze
+      RESULTS = ["tally.share", "tally.compensation", "end_tally"].freeze
 
       private
 
@@ -86,6 +86,9 @@ module VotingScheme
           end
         end
 
+        state[:compensations] = {}
+        state[:joint_compensations] = {}
+        state[:compensated] = 0
         state[:shares] = []
         state[:joint_shares] = build_questions_struct(1)
 
@@ -93,19 +96,24 @@ module VotingScheme
         append_content results
       end
 
-      def process_tally_message(message_identifier, _message, content)
-        raise RejectedMessage, "The owner_id doesn't match the sender trustee" if content[:owner_id] != message_identifier.author_id
-        raise RejectedMessage, "The trustee already sent their share" if state[:shares].include?(content[:owner_id])
+      def process_tally_message(message_identifier, message, content)
+        if message_identifier.subtype == "missing_trustee"
+          state[:compensations][message[:trustee_id]] = []
+          state[:joint_compensations][message[:trustee_id]] = build_questions_struct(1)
 
-        state[:shares] << content[:owner_id]
-
-        content[:contests].each do |question, answers|
-          answers.each do |answer, share|
-            state[:joint_shares][question][answer] *= share
-          end
+          return
         end
 
-        return unless state[:shares].count == election.trustees.count
+        raise RejectedMessage, "The owner_id doesn't match the sender trustee" if content[:owner_id] != message_identifier.author_id
+
+        case message_identifier.subtype
+        when "share"
+          process_tally_share_message content
+        when "compensation"
+          process_compensation_message content
+        end
+
+        return unless state[:shares].count + state[:compensated] == election.trustees.count
 
         results = build_questions_struct(0)
         state[:joint_shares].each do |question, answers|
@@ -115,6 +123,39 @@ module VotingScheme
         end
 
         emit_response "end_tally", results: results
+      end
+
+      def process_tally_share_message(content)
+        raise RejectedMessage, "The trustee already sent their share" if state[:shares].include?(content[:owner_id])
+
+        state[:shares] << content[:owner_id]
+
+        content[:contests].each do |question, answers|
+          answers.each do |answer, share|
+            state[:joint_shares][question][answer] *= share
+          end
+        end
+      end
+
+      def process_compensation_message(content)
+        compensation = state[:compensations][content[:trustee_id]]
+        raise RejectedMessage, "The trustee already sent their compensation for #{content[:trustee_id]}" if compensation.include?(content[:owner_id])
+
+        compensation << content[:owner_id]
+        content[:contests].each do |question, answers|
+          answers.each do |answer, value|
+            state[:joint_compensations][content[:trustee_id]][question][answer] *= value
+          end
+        end
+
+        return unless state[:compensations].count + state[:compensations][content[:trustee_id]].count == election.trustees.count
+
+        state[:joint_compensations][content[:trustee_id]].each do |question, answers|
+          answers.each do |answer, value|
+            state[:joint_shares][question][answer] *= (value.round**(1.0 / state[:shares].count)).round
+          end
+        end
+        state[:compensated] += 1
       end
 
       def build_questions_struct(initial_value)