+++ bulletin_board/server/spec/commands/report_missing_trustee_spec.rb
@@ -3,24 +3,20 @@
 require "rails_helper"
 require "./spec/commands/shared/log_entry_validations"
 
-RSpec.describe StartTally do
+RSpec.describe ReportMissingTrustee do
   subject { described_class.call(client, message_id, signed_data) }
 
   include_context "with a signed message"
 
   let!(:election) { create(:election, election_status) }
-  let(:election_status) { :vote_ended }
+  let(:election_status) { :tally }
   let(:client) { Authority.first }
-  let(:message_type) { :start_tally_message }
+  let(:message_type) { :report_missing_trustee_message }
   let(:message_params) { { election: election } }
 
-  shared_examples "starting the tally fails" do
+  shared_examples "reporting the missing trustee fails" do
     it "doesn't create a log entry" do
       expect { subject }.not_to change(LogEntry, :count)
-    end
-
-    it "doesn't change the election status" do
-      expect { subject }.not_to(change { Election.last.status })
     end
   end
 
@@ -28,24 +24,20 @@ RSpec.describe StartTally do
     expect { subject }.to broadcast(:ok)
   end
 
-  it "creates the log entry for the message and another for the response" do
-    expect { subject }.to change(LogEntry, :count).by(2)
+  it "creates the log entry for the message" do
+    expect { subject }.to change(LogEntry, :count).by(1)
   end
 
   it "persists the new state for the voting scheme" do
     expect { subject }.to(change { Election.last.voting_scheme_state })
   end
 
-  it "change the election status to tally" do
-    expect { subject }.to change { Election.last.status }.from("vote_ended").to("tally")
-  end
+  it_behaves_like "with an invalid signed data", "reporting the missing trustee fails"
 
-  it_behaves_like "with an invalid signed data", "starting the tally fails"
+  context "when the election status is not tally" do
+    let(:election_status) { :vote_ended }
 
-  context "when the election status is not vote_ended" do
-    let(:election_status) { :vote }
-
-    it_behaves_like "starting the tally fails"
+    it_behaves_like "reporting the missing trustee fails"
 
     it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "The election is not in the right step")
@@ -56,7 +48,7 @@ RSpec.describe StartTally do
     let(:client) { create(:authority, private_key: private_key) }
     let(:private_key) { generate(:private_key) }
 
-    it_behaves_like "starting the tally fails"
+    it_behaves_like "reporting the missing trustee fails"
 
     it "broadcast invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
@@ -67,7 +59,7 @@ RSpec.describe StartTally do
     let(:client) { Trustee.first }
     let(:private_key) { Test::PrivateKeys.trustees_private_keys.first }
 
-    it_behaves_like "starting the tally fails"
+    it_behaves_like "reporting the missing trustee fails"
 
     it "broadcast invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
@@ -77,7 +69,7 @@ RSpec.describe StartTally do
   context "when the message author is not the right authority" do
     let(:extra_message_params) { { authority: create(:authority) } }
 
-    it_behaves_like "starting the tally fails"
+    it_behaves_like "reporting the missing trustee fails"
 
     it "broadcast invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid message author")