+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+
+- The bulletin board client now includes a `report_missing_trustee` method to report the absence of a trustee during the Tally phase.
+
 ## [0.21.2] - 2021-05-28
 
 ## [0.21.1] - 2021-04-23